New changes in HTTP classes. HTTPUri class needs more changes, after adding Host member in HTTPData class.

+++ src/protocols/rest/HTTPSerializer.cpp
@@ -295,7 +295,7 @@ namespace eprosima
 
 		HTTPSerializer& HTTPSerializer::serialize(HTTPParameters &params, size_t numElements)
 		{
-			for(unsigned int i=0; i < numElements; ++i){
+			for(size_t i=0; i < numElements; ++i){
 				HTTPSerializer::serialize(params.get_params().at(i));
 			}
 			return *this;
@@ -305,7 +305,7 @@ namespace eprosima
 		{
 			HTTPParam param;
 			std::string parametersData = "";
-			for(unsigned int i=0; i < numElements; ++i){
+			for(size_t i=0; i < numElements; ++i){
 				HTTPSerializer::deserialize(param);
 				params.addParam(param);
 			}
